// src/types/enums/PokemonTypesEnum.ts
export enum PokemonTypesEnum {
    BUG = "bug",
    DARK = "dark",
    DRAGON = "dragon",
    ELECTRIC = "electric",
    FAIRY = "fairy",
    FIGHTING = "fighting",
    FIRE = "fire",
    FLYING = "flying",
    GHOST = "ghost",
    GROUND = "ground",
    GRASS = "grass",
    ICE = "ice",
    NORMAL = "normal",
    POISON = "poison",
    PSYCHIC = "psychic",
    ROCK = "rock",
    STEEL = "steel",
    WATER = "water"
}